Adding validation for options
