+++ src/Viserio/Component/Contracts/OptionsResolver/RequiresValidatedConfig.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Component\Contracts\OptionsResolver;
+
+interface RequiresValidatedConfig
+{
+    /**
+     * Returns a list of option validator callables which must be available in getMandatoryOptions.
+     *
+     * @return array
+     */
+    public function getOptionValidators(): array;
+}
