+++ src/Viserio/Component/Cookie/Providers/CookieServiceProvider.php
@@ -64,13 +64,29 @@ class CookieServiceProvider implements
 
     public static function createCookieJar(ContainerInterface $container): CookieJar
     {
-        $options = self::resolveOptions($container);
+        $options = self::resolve($container);
 
         return (new CookieJar())->setDefaultPathAndDomain(
             $options['path'],
             $options['domain'],
             $options['secure']
         );
+    }
+
+    protected static function resolveConfiguration($data)
+    {
+        if (is_iterable($data)) {
+            return $data;
+        } elseif ($data instanceof ContainerInterface) {
+            if ($data->has(RepositoryContract::class)) {
+                return $data->get(RepositoryContract::class);
+            } elseif ($data->has('config')) {
+                return $data->get('config');
+            } elseif ($data->has('options')) {
+                return $data->get('options');
+            }
+        }
+        throw new RuntimeException('No configuration found.');
     }
 
     /**
