+++ src/Viserio/Component/Cookie/Tests/Providers/CookieServiceProviderTest.php
@@ -6,14 +6,12 @@ use PHPUnit\Framework\TestCase;
 use Viserio\Component\Container\Container;
 use Viserio\Component\Cookie\CookieJar;
 use Viserio\Component\Cookie\Providers\CookieServiceProvider;
-use Viserio\Component\OptionsResolver\Providers\OptionsResolverServiceProvider;
 
 class CookieServiceProviderTest extends TestCase
 {
     public function testProvider()
     {
         $container = new Container();
-        $container->register(new OptionsResolverServiceProvider());
         $container->register(new CookieServiceProvider());
 
         $container->instance('config', [
