+++ src/Viserio/Component/OptionsResolver/Traits/StaticOptionsResolverTrait.php
@@ -17,6 +17,8 @@ use Viserio\Component\Contracts\OptionsResolver\RequiresMandatoryOptions as Requ
 
 trait StaticOptionsResolverTrait
 {
+    use AbstractOptionsResolverTrait;
+
     /**
      * Returns options based on getDimensions() like [vendor][package] if class implements RequiresComponentConfig
      * and can perform mandatory option checks if class implements RequiresMandatoryOptions. If the
@@ -25,8 +27,8 @@ trait StaticOptionsResolverTrait
      *
      * The RequiresConfigId interface is supported.
      *
-     * @param null|string $configId Config name, must be provided if factory uses RequiresConfigId interface
      * @param mixed       $config
+     * @param null|string $configId Config name, must be provided if factory uses RequiresConfigId interface
      *
      * @throws \InvalidArgumentException                                                                If the $configId parameter is provided but factory does not support it
      * @throws \Viserio\Component\Contracts\OptionsResolver\Exceptions\UnexpectedValueException         If the $config parameter has the wrong type
@@ -35,50 +37,13 @@ trait StaticOptionsResolverTrait
      *
      * @return array
      */
-    public static function resolveOptions($config, string $configId = null): array
+    protected static function resolve($config, string $configId = null): array
     {
-        $config      = self::resolveConfiguration($config);
-        $configClass = self::getConfigClass();
-        $dimensions  = [];
-
-        if ($configClass instanceof RequiresComponentConfigContract) {
-            $dimensions  = $configClass->getDimensions();
-            $dimensions  = $dimensions instanceof Iterator ? iterator_to_array($dimensions) : $dimensions;
-        }
-
-        if ($configClass instanceof RequiresConfigIdContract ||
-            $configClass instanceof RequiresComponentConfigIdContract
-        ) {
-            $dimensions[] = $configId;
-        } elseif ($configId !== null) {
-            throw new InvalidArgumentException(
-                sprintf('The factory [%s] does not support multiple instances.', get_class($configClass))
-            );
-        }
-
-        // get configuration for provided dimensions
-        $config = self::getConfigurationDimensions($dimensions, $config, $configClass, $configId);
-
-        if ((array) $config !== $config &&
-            ! $config instanceof ArrayAccess &&
-            $configClass instanceof RequiresComponentConfigIdContract
-        ) {
-            throw new UnexpectedValueException($configClass->getDimensions());
-        }
-
-        if ($configClass instanceof RequiresMandatoryOptionsContract) {
-            self::checkMandatoryOptions($configClass->getMandatoryOptions(), $config);
-        }
-
-        if ($configClass instanceof ProvidesDefaultOptionsContract) {
-            $options = $configClass->getDefaultOptions();
-            $config  = array_replace_recursive(
-                $options instanceof Iterator ? iterator_to_array($options) : (array) $options,
-                (array) $config
-            );
-        }
-
-        return (array) $config;
+        return self::resolveOptions(
+            self::resolveConfiguration($config),
+            self::getConfigClass(),
+            $configId
+        );
     }
 
     /**
@@ -90,85 +55,12 @@ trait StaticOptionsResolverTrait
      *
      * @return array|\ArrayAccess
      */
-    abstract protected static function resolveConfiguration($data);
+    abstract static protected function resolveConfiguration($data);
 
     /**
      * The configurable class.
      *
      * @return \Viserio\Component\Contracts\OptionsResolver\RequiresConfig
      */
-    abstract protected static function getConfigClass(): RequiresConfigContract;
-
-    /**
-     * Checks if a mandatory param is missing, supports recursion.
-     *
-     * @param iterable $mandatoryOptions
-     * @param iterable $config
-     *
-     * @throws \Viserio\Component\Contracts\OptionsResolver\Exceptions\MandatoryOptionNotFoundException
-     *
-     * @return void
-     */
-    private static function checkMandatoryOptions(iterable $mandatoryOptions, iterable $config): void
-    {
-        foreach ($mandatoryOptions as $key => $mandatoryOption) {
-            $useRecursion = ! is_scalar($mandatoryOption);
-
-            if (! $useRecursion && isset($config[$mandatoryOption])) {
-                continue;
-            }
-
-            if ($useRecursion && isset($config[$key])) {
-                self::checkMandatoryOptions($mandatoryOption, $config[$key]);
-
-                return;
-            }
-
-            $configClass = self::getConfigClass();
-
-            throw new MandatoryOptionNotFoundException(
-                $configClass instanceof RequiresComponentConfigContract ? $configClass->getDimensions() : [],
-                $useRecursion ? $key : $mandatoryOption
-            );
-        }
-    }
-
-    /**
-     * Get configuration for provided dimensions.
-     *
-     * @param iterable                                                    $dimensions
-     * @param iterable                                                    $config
-     * @param \Viserio\Component\Contracts\OptionsResolver\RequiresConfig $configClass
-     * @param null|string                                                 $configId
-     *
-     * @throws \Viserio\Component\Contracts\OptionsResolver\Exceptions\OptionNotFoundException
-     *
-     * @return iterable|object
-     */
-    private static function getConfigurationDimensions(
-        iterable $dimensions,
-        $config,
-        RequiresConfigContract $configClass,
-        ?string $configId
-    ) {
-        foreach ($dimensions as $dimension) {
-            if ((array) $config !== $config && ! $config instanceof ArrayAccess) {
-                throw new UnexpectedValueException($dimensions, $dimension);
-            }
-
-            if (! isset($config[$dimension])) {
-                if (! $configClass instanceof RequiresMandatoryOptionsContract &&
-                    $configClass instanceof ProvidesDefaultOptionsContract
-                ) {
-                    break;
-                }
-
-                throw new OptionNotFoundException($configClass, $dimension, $configId);
-            }
-
-            $config = $config[$dimension];
-        }
-
-        return $config;
-    }
+    abstract static protected function getConfigClass(): RequiresConfigContract;
 }
